chore: move TestWorkflow mappers to common mapperstcl directory
