+++ cmd/kubectl-testkube/commands/testworkflows/get.go
@@ -10,7 +10,7 @@ import (
 	"github.com/kubeshop/testkube/cmd/kubectl-testkube/commands/common"
 	"github.com/kubeshop/testkube/cmd/kubectl-testkube/commands/common/render"
 	"github.com/kubeshop/testkube/cmd/kubectl-testkube/commands/testworkflows/renderer"
-	"github.com/kubeshop/testkube/pkg/tcl/workflowstcl/mappers"
+	"github.com/kubeshop/testkube/pkg/tcl/mapperstcl/mappers"
 	"github.com/kubeshop/testkube/pkg/ui"
 )
