+++ pkg/tcl/mapperstcl/mappers/kube_openapi.go
